backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Celery
redis==5.0.1
celery==5.3.6
flower==2.0.1

# AI & ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
qdrant-client==1.7.3
tiktoken==0.5.2

# Document Generation
python-docx==1.1.0
jinja2==3.1.3
pdfkit==1.0.0
openpyxl==3.1.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Utils
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3
pendulum==3.0.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.14
mypy==1.8.0

# Email
fastapi-mail==1.4.1

# Rate Limiting
slowapi==0.1.9

